Handling continuous scroll key input.
Add impulse checking the current velocity value.
Without this patch, keeping pressing scroll key, it doessn't stop
properly. This patch make to set a upper limit to velocity value.

Add a function for distance instead of impluse. It will help to move a
specific lines. But it will ignore the air_drag value. It is hard to
calculate the impulse value to move a certain distance with air_drag.